Share the SQL text helpers between the capacity test files
The split left both files with the same occurrences and whitespace
helpers; they live in #test-utils/sql-text.ts now.
